Switch to Ubuntu rolling (currently 17.10 artful)
17.04 zesty was an awkward short-living release (like all non-LTS
rolling releases). It was already out of the "rolling" designation by
the time we chose it. And we only chose it because of new compiler
warnings from GCC 7. Now that these have all been fixed, let's keep
rolling forward! [Fixes #39]

This commit also integrates testing of the image right into the process
of "docker build": no more excuses for not running tests! [Closes #38]