Bypass container layer for cluster data directories
AUFS and overlay2 directories are slow to traverse. I'm not suggesting
we should not do away with the questionable way we drop AO tables, but
using the host FS like ext4 and xfs buys us a lot of speed in the near
term.
This reduces 25% of ICG time for ORCA, and ~40% of ICG time for planner:

optimizer|before/after this patch|time
---|---|---
off|before|27m6.137s
off|after|16m9.845s
on|before|31m12.139s
on|after|23m15.333s