Use latest shellcheck (from Travis)
Travis seems to always provide the latest shellcheck from Hackage /
Cabal, which is great. Debian's shellcheck is usually a half a year
behind.  Even though /usr/bin/shellcheck doesn't seem to conflict with
the shellcheck from Travis (because of ordering in PATH), this confuses
me.
Remove apt installation of shellcheck.